refactored two wiki created files
